2-Projects/量化交易/量化交易1/Ashare复盘multi-agents/next_day_expectation_agent/requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pyyaml>=6.0
pydantic>=2.0.0

# Data sources
akshare>=1.12.0
requests>=2.31.0

# LLM integration
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Utilities
python-dotenv>=1.0.0
jsonschema>=4.19.0
